backend/requirements.txt
numpy
pandas
Flask
Flask-CORS
scikit-learn
requests
Pillow
markupsafe
python-dotenv
google-generativeai
openai
PyPDF2
pytesseract
pdf2image